+++ core-bundle/src/Resources/contao/library/Contao/File.php
@@ -763,6 +763,8 @@ class File extends \System
 	public function sendToBrowser($filename='')
 	{
 		$response = new BinaryFileResponse(TL_ROOT . '/' . $this->strFile);
+		$response->setPrivate(); // public by default
+		$response->setAutoEtag();
 
 		$response->setContentDisposition
 		(
@@ -772,9 +774,6 @@ class File extends \System
 		);
 
 		$response->headers->addCacheControlDirective('must-revalidate');
-		$response->headers->addCacheControlDirective('post-check', 0);
-		$response->headers->addCacheControlDirective('pre-check', 0);
-
 		$response->headers->set('Connection', 'close');
 
 		throw new ResponseException($response);